Move cdi_2_surface out of processing.py
