+++ geoapps/processing/__init__.py
@@ -1,5 +1,4 @@
 from .processing import (
-    CDICurve2Surface,
     calculator,
     coordinate_transformation_widget,
 )
@@ -7,3 +6,4 @@ from .contours import ContourValues
 from .em_profiling import EMLineProfiler
 from .edge_detection import EdgeDetectionApp
 from .data_interpolation import DataInterpolation
+from .cdi_surface import CDICurve2Surface
