+++ geoapps/processing/processing.py
@@ -103,410 +103,6 @@ def calculator(h5file):
     )
 
 
-class CDICurve2Surface(ObjectDataSelection):
-    """
-    Application for the conversion of conductivity/depth curves to
-    a pseudo 3D conductivity model on surface.
-    """
-
-    defaults = {
-        "add_groups": True,
-        "select_multiple": True,
-        "h5file": "../../assets/FlinFlon.geoh5",
-        "objects": "CDI_VTEM_model",
-        "data": ["COND"],
-        "max_distance": 100,
-        "max_depth": 400,
-        "elevations": {"data": "ELEV"},
-        "lines": {"data": "Line"},
-    }
-
-    def __init__(self, **kwargs):
-        kwargs = self.apply_defaults(**kwargs)
-        self._lines = ObjectDataSelection(add_groups=True, find_value=["line"])
-        self._topography = TopographyOptions()
-        self._elevations = ObjectDataSelection(add_groups=True)
-        self._z_option = widgets.RadioButtons(
-            options=["elevation", "depth"], description="Layers reference:",
-        )
-        self._max_depth = widgets.FloatText(description="Max depth (m):",)
-        self._max_distance = widgets.FloatText(
-            value=50, description="Max distance (m):",
-        )
-        self._tolerance = widgets.FloatText(value=1, description="Tolerance (m):")
-        self._save_name = widgets.Text("CDI_", description="Name: ")
-        self._convert = widgets.ToggleButton(
-            description="Convert >>", button_style="success"
-        )
-
-        super().__init__(**kwargs)
-
-        if "lines" in kwargs.keys():
-            self.lines.__populate__(**kwargs["lines"])
-        if "topography" in kwargs.keys():
-            self.topography.__populate__(**kwargs["topography"])
-        if "elevations" in kwargs.keys():
-            self.elevations.__populate__(**kwargs["elevations"])
-
-        self.lines.data.description = "Line field:"
-        self.elevations.data.description = "Elevations:"
-        self.data.description = "Model fields: "
-
-        def z_options_change(_):
-            self.z_options_change()
-
-        self.z_option.observe(z_options_change)
-        self.depth_panel = widgets.HBox([self.z_option, self.elevations.data])
-
-        def convert_trigger(_):
-            self.convert_trigger()
-
-        self.convert.observe(convert_trigger)
-        self._widget = widgets.VBox(
-            [
-                self.project_panel,
-                widgets.VBox(
-                    [
-                        widgets.VBox(
-                            [
-                                self.widget,
-                                self.lines.data,
-                                self.depth_panel,
-                                widgets.Label("Triangulation"),
-                                self.max_depth,
-                                self.max_distance,
-                                self.tolerance,
-                            ]
-                        ),
-                    ]
-                ),
-                widgets.Label("Output"),
-                widgets.HBox([self.convert, self.save_name,]),
-            ]
-        )
-
-    def convert_trigger(self):
-
-        if self.convert.value:
-            if self.workspace.get_entity(self.objects.value):
-                curve = self.workspace.get_entity(self.objects.value)[0]
-                self.convert.value = False
-            else:
-                self.convert.button_style = "warning"
-                self.convert.value = False
-                return
-
-            lines_id = curve.get_data(self.lines.data.value)[0].values
-            lines = numpy.unique(lines_id).tolist()
-
-            if self.z_option.value == "depth":
-                if self.topography.options_button.value == "Object":
-
-                    topo_obj = self.workspace.get_entity(self.topography.objects.value)[
-                        0
-                    ]
-
-                    if hasattr(topo_obj, "centroids"):
-                        vertices = topo_obj.centroids.copy()
-                    else:
-                        vertices = topo_obj.vertices.copy()
-
-                    topo_xy = vertices[:, :2]
-
-                    if self.topography.value.value == "Vertices":
-                        topo_z = vertices[:, 2]
-                    else:
-                        topo_z = topo_obj.get_data(self.topography.value.value)[
-                            0
-                        ].values
-
-                else:
-                    topo_xy = curve.vertices[:, :2].copy()
-
-                    if self.topography.options_button.value == "Constant":
-                        topo_z = (
-                            numpy.ones_like(curve.vertices[:, 2])
-                            * self.topography.constant.value
-                        )
-                    else:
-                        topo_z = (
-                            numpy.ones_like(curve.vertices[:, 2])
-                            + self.topography.offset.value
-                        )
-
-                surf = Delaunay(topo_xy)
-                topo = LinearNDInterpolator(surf, topo_z)
-                tree_topo = cKDTree(topo_xy)
-
-            model_vertices = []
-            model_cells = []
-            model_count = 0
-            locations = curve.vertices
-            model = []
-            line_ids = []
-            for line in lines:
-
-                line_ind = numpy.where(lines_id == line)[0]
-
-                n_sounding = len(line_ind)
-                if n_sounding < 2:
-                    continue
-
-                xyz = locations[line_ind, :]
-
-                # Create a 2D mesh to store the results
-                if numpy.std(xyz[:, 1]) > numpy.std(xyz[:, 0]):
-                    order = numpy.argsort(xyz[:, 1])
-                else:
-                    order = numpy.argsort(xyz[:, 0])
-
-                X, Y, Z, M = [], [], [], []
-                # Stack the z-coordinates and model
-                nZ = 0
-                for ind, z_prop in enumerate(
-                    curve.get_property_group(self.elevations.data.value).properties,
-                ):
-                    nZ += 1
-                    z_vals = self.workspace.get_entity(z_prop)[0].values[line_ind]
-
-                    m_vals = []
-                    for m in self.data.value:
-                        prop = curve.get_property_group(m).properties[ind]
-                        m_vals.append(
-                            self.workspace.get_entity(prop)[0].values[line_ind]
-                        )
-
-                    m_vals = numpy.vstack(m_vals).T
-                    keep = (
-                        (z_vals > 1e-38)
-                        * (z_vals < 2e-38)
-                        * numpy.any((m_vals > 1e-38) * (m_vals < 2e-38), axis=1)
-                    ) == False
-                    keep[numpy.isnan(z_vals)] = False
-                    keep[numpy.any(numpy.isnan(m_vals), axis=1)] = False
-
-                    X.append(xyz[:, 0][order][keep])
-                    Y.append(xyz[:, 1][order][keep])
-
-                    if self.z_option.value == "depth":
-                        z_topo = topo(xyz[:, 0][order][keep], xyz[:, 1][order][keep])
-
-                        nan_z = numpy.isnan(z_topo)
-                        if numpy.any(nan_z):
-                            _, ii = tree_topo.query(xyz[:, :2][order][keep][nan_z])
-                            z_topo[nan_z] = topo_z[ii]
-
-                        Z.append(z_topo + z_vals[order][keep])
-
-                    else:
-                        Z.append(z_vals[order][keep])
-
-                    M.append(m_vals[order, :][keep, :])
-
-                    if ind == 0:
-
-                        x_loc = xyz[:, 0][order][keep]
-                        y_loc = xyz[:, 1][order][keep]
-                        z_loc = Z[0]
-
-                X = numpy.hstack(X)
-                Y = numpy.hstack(Y)
-                Z = numpy.hstack(Z)
-                model.append(numpy.vstack(M))
-                line_ids.append(numpy.ones_like(Z.ravel()) * line)
-
-                if numpy.std(y_loc) > numpy.std(x_loc):
-                    tri2D = Delaunay(numpy.c_[numpy.ravel(Y), numpy.ravel(Z)])
-                    dist = numpy.ravel(Y)
-                    topo_top = interp1d(y_loc, z_loc)
-                else:
-                    tri2D = Delaunay(numpy.c_[numpy.ravel(X), numpy.ravel(Z)])
-                    dist = numpy.ravel(X)
-                    topo_top = interp1d(x_loc, z_loc)
-
-                    # Remove triangles beyond surface edges
-                indx = numpy.ones(tri2D.simplices.shape[0], dtype=bool)
-                for ii in range(3):
-                    x = numpy.mean(
-                        numpy.c_[
-                            dist[tri2D.simplices[:, ii]],
-                            dist[tri2D.simplices[:, ii - 1]],
-                        ],
-                        axis=1,
-                    )
-                    z = numpy.mean(
-                        numpy.c_[
-                            Z[tri2D.simplices[:, ii]], Z[tri2D.simplices[:, ii - 1]]
-                        ],
-                        axis=1,
-                    )
-
-                    length = numpy.linalg.norm(
-                        tri2D.points[tri2D.simplices[:, ii], :]
-                        - tri2D.points[tri2D.simplices[:, ii - 1], :],
-                        axis=1,
-                    )
-
-                    indx *= (
-                        (z <= (topo_top(x) + self.tolerance.value))
-                        * (
-                            z
-                            >= (
-                                topo_top(x)
-                                - self.max_depth.value
-                                - self.tolerance.value
-                            )
-                        )
-                        * (length < self.max_distance.value)
-                    )
-
-                # Remove the simplices too long
-                tri2D.simplices = tri2D.simplices[indx, :]
-                tri2D.vertices = tri2D.vertices[indx, :]
-
-                temp = numpy.arange(int(nZ * n_sounding)).reshape(
-                    (nZ, n_sounding), order="F"
-                )
-                model_vertices.append(
-                    numpy.c_[numpy.ravel(X), numpy.ravel(Y), numpy.ravel(Z)]
-                )
-                model_cells.append(tri2D.simplices + model_count)
-
-                model_count += tri2D.points.shape[0]
-
-            surface = Surface.create(
-                self.workspace,
-                name=self.save_name.value,
-                vertices=numpy.vstack(model_vertices),
-                cells=numpy.vstack(model_cells),
-            )
-
-            surface.add_data(
-                {"Line": {"values": numpy.hstack(line_ids)},}
-            )
-
-            models = numpy.vstack(model)
-            for ind, field in enumerate(self.data.value):
-
-                surface.add_data(
-                    {field: {"values": models[:, ind]},}
-                )
-
-    def z_options_change(self):
-        if self.z_option.value == "depth":
-            self.elevations.data.description = "Depth:"
-            self.depth_panel.children = [
-                self.z_option,
-                widgets.VBox(
-                    [
-                        self.elevations.data,
-                        widgets.Label("Topography"),
-                        self.topography.widget,
-                    ]
-                ),
-            ]
-        else:
-            self.elevations.data.description = "Elevation:"
-            self.depth_panel.children = [self.z_option, self.elevations.data]
-
-    @property
-    def lines(self):
-        """
-            Line field options
-        """
-        return self._lines
-
-    @property
-    def topography(self):
-        """
-            TopographyOptions()
-        """
-        return self._topography
-
-    @property
-    def elevations(self):
-        """
-            ObjectDataSelection()
-        """
-        return self._elevations
-
-    @elevations.setter
-    def elevations(self, value):
-        assert isinstance(
-            value, ObjectDataSelection
-        ), f"elevations must be an object of type {ObjectDataSelection}"
-        self._elevations = value
-
-    @property
-    def z_option(self):
-        """
-            widgets.RadioButtons()
-        """
-        return self._z_option
-
-    @property
-    def max_depth(self):
-        """
-            widgets.FloatText()
-        """
-        return self._max_depth
-
-    @property
-    def max_distance(self):
-        """
-            widgets.FloatText()
-        """
-        return self._max_distance
-
-    @property
-    def tolerance(self):
-        """
-            widgets.FloatText()
-        """
-        return self._tolerance
-
-    @property
-    def save_name(self):
-        """
-            widgets.Text()
-        """
-        return self._save_name
-
-    @property
-    def convert(self):
-        """
-            widgets.ToggleButton()
-        """
-        return self._convert
-
-    @property
-    def workspace(self):
-        """
-        Target geoh5py workspace
-        """
-        if (
-            getattr(self, "_workspace", None) is None
-            and getattr(self, "_h5file", None) is not None
-        ):
-            self.workspace = Workspace(self.h5file)
-        return self._workspace
-
-    @workspace.setter
-    def workspace(self, workspace):
-        assert isinstance(workspace, Workspace), f"Workspace must of class {Workspace}"
-        self._workspace = workspace
-        self._h5file = workspace.h5file
-
-        # Refresh the list of objects
-        self.update_objects_list()
-
-        self.lines.workspace = workspace
-        self.lines.objects = self.objects
-
-        self.elevations.workspace = workspace
-        self.elevations.objects = self.objects
-
-
 def coordinate_transformation_widget(
     h5file, plot=False, epsg_in=None, epsg_out=None, object_names=[]
 ):
